Add animation on HUD_H Change data helpSystem to enable help after 30sec of inactivation of player to do test button Try reset slider value setting

+++ Assets/Components/MaxFontSize.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class MaxFontSize : MonoBehaviour {
     // Advice: FYFY component aims to contain only public members (according to Entity-Component-System paradigm).
-    public int maxSize; //Valeur renseignée dans le max size jusqu'où le texte peut grossir au max
+    public float maxSize; //Valeur renseignée dans le max size jusqu'où le texte peut grossir au max
     [HideInInspector]
     public float defaultMaxSize; //Valeur renseignée dans le max size qui a permis de définir la taille de la police de départ
